requirements.txt
discord.py>=2.0.0
confluent-kafka>=2.0.0
pydantic>=2.0.0
pytest>=7.0.0
mypy>=1.0.0
flake8>=6.0.0